changed max redirects to 3 for rrb plugin
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@6554 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab

+++ nexus/nexus-core-plugins/nexus-rrb-plugin/src/main/java/org/sonatype/nexus/plugins/rrb/MavenRepositoryReader.java
@@ -204,7 +204,9 @@ public class MavenRepositoryReader
         }
 
         method.getParams().setParameter( HttpMethodParams.RETRY_HANDLER, new DefaultHttpMethodRetryHandler( 3, false ) );
-        client.getParams().setParameter( "http.protocol.max-redirects", new Integer( 1 ) );
+        
+        // allow a couple redirects
+        client.getParams().setParameter( "http.protocol.max-redirects", new Integer( 3 ) );
 
         StringBuilder result = new StringBuilder();
